Rework graph structure to represent nodes as integer ids
This is a first step towards decoupling the internal data structures
from lots of direct references to little node objects. Instead of
storing graph nodes directly in the adjacency list, we now store ids.

As a consequence, all the API quirks and specs dealing with equality
based on identity have been removed and the `NodeProxy` is used
exclusively to represent in-memory node instances.