Support implicit node or collection in pipeline processing
Assumes that each step in the pipeline chain can either be passed a
single node or a collection of nodes. This fixes the failing tests which
roughly explained what the expected behaviour should be.

However, this is all using eager/imperative evaluation and as a
consequence, the pipeline semantics are limited to forced looping over
the entire collection at each step.