Extract mutator API into separate module and introduce a mutable graph
There’s now a mutable version of the core graph class which supports the
entire range of add, set and create methods. Implementation is now being
shared between the graph builder and the mutable graph class.

This suggests a more modular design for the entire API might be
possible, without too much heavy lifting in the current codebase.